requirements.txt
flask
python-dotenv
apscheduler
landingai-ade
pydantic
psycopg[binary]
stripe
google-generativeai
anthropic
